LYWSD03MMC_monitoring_adafruitio/requirements.txt
Adafruit-Blinka==6.14.1
adafruit-blinka-bleio==2.2.0
adafruit-circuitpython-ble==8.1.0
adafruit-circuitpython-ble-lywsd03mmc==1.0.2
adafruit-io==2.5.0
Adafruit-PlatformDetect==3.15.3
Adafruit-PureIO==1.1.9
bleak==0.11.0
certifi==2021.10.8
charset-normalizer==2.0.6
dbus-next==0.2.3
idna==3.2
paho-mqtt==1.5.1
pyftdi==0.53.3
pyserial==3.5
python-dotenv==0.19.1
pyusb==1.2.1
requests==2.26.0
rpi-ws281x==4.3.0
RPi.GPIO==0.7.0
sysv-ipc==1.1.0
urllib3==1.26.7
